requirement.txt
albumentations==1.0.0
appdirs==1.4.4
ase==3.21.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autopep8==1.5.7
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.9.3
beautify==0.1.4
black==19.10b0
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
blis==0.7.4
Bottleneck==1.3.2
cached-property==1.5.2
catalogue==1.0.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.1
cycler==0.10.0
cymem==2.0.5
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dgl==0.6.0.post1
efficientnet-pytorch==0.6.3
entrypoints==0.3
fastai==1.0.60
fastcore==1.3.19
fastprogress==1.0.0
flake8 @ file:///tmp/build/80754af9/flake8_1620776156532/work
Flask==1.1.1
Flask-Cors==3.0.10
future==0.18.2
googledrivedownloader==0.4
gunicorn==20.0.4
h5py==3.2.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1615900449442/work
imutils==0.5.3
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206598566/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1614615642284/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isodate==0.6.0
itsdangerous==1.1.0
jedi @ file:///tmp/build/80754af9/jedi_1606932558467/work
Jinja2==2.11.1
joblib==1.0.1
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213308260/work
kiwisolver==1.1.0
llvmlite==0.36.0
lxml==4.6.2
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1594371495811/work
matplotlib==3.2.1
mccabe==0.6.1
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
murmurhash==1.0.5
mypy-extensions==0.4.3
nbconvert @ file:///tmp/build/80754af9/nbconvert_1594376811065/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1594379827080/work
numba==0.53.0
numexpr==2.7.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==4.2.0.32
opencv-python-headless==4.5.2.54
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.3
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
parso==0.7.0
pathspec==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1615057347459/work
plac==1.1.3
preshed==3.0.5
pretrainedmodels==0.7.4
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pyflakes @ file:///tmp/build/80754af9/pyflakes_1617200973297/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pymatreader==0.0.24
pyparsing==2.4.6
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141707582/work
PySide2==5.15.2
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-louvain==0.15
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
rdflib==5.0.0
regex @ file:///tmp/build/80754af9/regex_1616094021210/work
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy @ file:///tmp/build/80754af9/scipy_1614022789548/work
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@a6e1123983548be55d4d1320e0a2f5fd9174d4ac
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
Shapely==1.7.1
shiboken2==5.15.2
six==1.14.0
soupsieve==2.2
spacy==2.3.5
srsly==1.0.5
terminado==0.9.3
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
thinc==7.4.5
threadpoolctl==2.1.0
tifffile==2021.4.8
timm==0.3.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.8.0
torch-cluster==1.5.9
torch-geometric==1.6.3
torch-scatter==2.0.6
torch-sparse==0.6.9
torch-spline-conv==1.2.1
torchaudio==0.8.0a0+a751e1d
torchvision==0.9.0
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.59.0
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typed-ast @ file:///tmp/build/80754af9/typed-ast_1610484562261/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.4
wasabi==0.8.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.0
xmltodict==0.12.0
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
